frontend: Fix panic during unit tests in EmitCounter
EmitCounter was registering its collector to the default registerer
instead of the PrometheusEmitter registry, so unit tests that used
PrometheusEmitter would panic on the second test case for trying to
register a duplicate collector.

This was being caught and obfuscated, however, by MiddlewarePanic,
and would often manifest in unit test output as

    http: superfluous response.WriteHeader call from ...

when MiddlewarePanic attempted to call WriteInternalServerError.